refactor: cognito auth server actions
